+++ app/src/app_abacus_plot.py
@@ -51,12 +51,12 @@ def generate_abacus_plot(ns_changes, df_join, min_samples, df_haplotypes_set, ge
     # Filter QC fail and missing samples
     df_samples_with_ns_changes = df_join.loc[
         (df_join['QC pass'] == True) &
-        (df_join["Exclusion reason"] == 'Analysis_set') &
-        (df_join["HaploAtlas exclusion reason"] == "Analysis_set")
+        (df_join["Exclusion reason"].isin(['Analysis_set', 'Unassigned_population'])) &
+        (df_join["HaploAtlas exclusion reason"].isin(['Analysis_set', 'Unassigned_population']))
     ].copy()
     
     df_samples_with_ns_changes.loc[df_samples_with_ns_changes['Country'] == 'Democratic Republic of the Congo', ['Country']] = 'DRC'
-    df_samples_with_ns_changes.loc[df_samples_with_ns_changes.ns_changes == "", "ns_changes"] = "3D7 REF"
+    df_samples_with_ns_changes.loc[df_samples_with_ns_changes.ns_changes == "", "ns_changes"] = "P01"
 
     df_samples_with_ns_changes["Admin level 1"] = df_samples_with_ns_changes["Admin level 1"].fillna("(Unspecified)")